extracts mapping into separate module and fixes linting
