+++ snmp-discovery/snmp/snmp_test.go
@@ -13,6 +13,7 @@ import (
 	"github.com/stretchr/testify/mock"
 
 	"github.com/netboxlabs/orb-discovery/snmp-discovery/config"
+	"github.com/netboxlabs/orb-discovery/snmp-discovery/mapping"
 	"github.com/netboxlabs/orb-discovery/snmp-discovery/snmp"
 )
 
@@ -139,47 +140,9 @@ func (m *MockSNMP) Close() error {
 }
 
 // Walk implements Walker interface
-func (m *MockSNMP) Walk(oid string) (snmp.ObjectIDValueMap, error) {
+func (m *MockSNMP) Walk(oid string) (mapping.ObjectIDValueMap, error) {
 	args := m.Called(oid)
 	return nil, args.Error(1)
-}
-
-func TestMapObjectIDsToEntity(t *testing.T) {
-	mapper := snmp.NewObjectIDMapper([]config.MappingEntry{
-		{
-			OID:    "1.3.6.1.2.1.4.20.1.1",
-			Entity: "ipAddress",
-			Field:  "address",
-		},
-	})
-	objectIDs := snmp.ObjectIDValueMap{
-		"1.3.6.1.2.1.4.20.1.1": "192.168.1.1",
-	}
-
-	entities := mapper.MapObjectIDsToEntity(objectIDs)
-
-	assert.Len(t, entities, 1)
-	ipEntity, ok := entities[0].(*diode.IPAddress)
-	assert.True(t, ok)
-	assert.Equal(t, "192.168.1.1/32", *ipEntity.Address)
-}
-
-func TestObjectIDs(t *testing.T) {
-	mapper := snmp.NewObjectIDMapper([]config.MappingEntry{
-		{
-			OID:    "1.3.6.1.2.1.4.20.1.1",
-			Entity: "ipAddress",
-			Field:  "address",
-		},
-	})
-
-	expectedObjectIDs := []string{
-		"1.3.6.1.2.1.4.20.1.1",
-	}
-
-	objectIDs := mapper.ObjectIDs()
-
-	assert.ElementsMatch(t, expectedObjectIDs, objectIDs)
 }
 
 func TestNewClient(t *testing.T) {
